Moved likes system to idea_forum app from the overridden machina forum_conversation app
